copr: move logical rows to separate mod
