+++ components/tidb_query_datatype/src/codec/data_type/mod.rs
@@ -5,8 +5,10 @@ mod chunked_vec_bytes;
 mod chunked_vec_common;
 mod chunked_vec_json;
 mod chunked_vec_sized;
+mod logical_rows;
 mod scalar;
 mod vector;
+pub use logical_rows::{LogicalRows, BATCH_MAX_SIZE};
 
 // Concrete eval types without a nullable wrapper.
 pub type Int = i64;
